requirements.txt
google-generativeai>=0.3.0
python-dateutil>=2.8.0
requests>=2.28.0
langchain
langchain-google-genai
python-telegram-bot
pyTelegramBotAPI
python-dotenv
duckduckgo-search